requirements.txt
decord==0.6.0
einops==0.8.0
fairscale==0.4.13
fvcore==0.1.5.post20221221
numpy==1.22.0
PyYAML==6.0.2
thop==0.1.1.post2209072238
timm==1.0.12
torch==1.13.1+cu116
torchvision==0.14.1+cu116
transformers==4.36.2
